node ausgegliedert
